requirements.txt
gunicorn
mysql-connector-python
Flask
flask-login
flask-WTF
wtforms
flask-bootstrap
email_validator
argon2-cffi